revert(src): reverted API break: do not remove old backlight module support.
This allows to release a new Clightd version with the new "emulated backlight feature"
without touching Clight at all.
